# docs/openapi/api-v1.openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "operationId": "AuthController_register_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRegisterBodyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterUserResponseSchema"
                }
              }
            }
          },
          "400": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsBodySchema"
                }
              }
            }
          },
          "409": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsBodySchema"
                }
              }
            }
          }
        },
        "summary": "Register a new user",
        "tags": ["auth"]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLoginBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginUserResponseSchema"
                }
              }
            }
          },
          "400": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsBodySchema"
                }
              }
            }
          },
          "401": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsBodySchema"
                }
              }
            }
          },
          "403": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsBodySchema"
                }
              }
            }
          },
          "429": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsBodySchema"
                }
              }
            }
          }
        },
        "summary": "Login with email and password",
        "tags": ["auth"]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRefreshBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshSessionResponseSchema"
                }
              }
            }
          },
          "400": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsBodySchema"
                }
              }
            }
          },
          "401": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsBodySchema"
                }
              }
            }
          },
          "403": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsBodySchema"
                }
              }
            }
          }
        },
        "summary": "Rotate refresh token and issue new access token",
        "tags": ["auth"]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRefreshBodyDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsBodySchema"
                }
              }
            }
          },
          "401": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsBodySchema"
                }
              }
            }
          }
        },
        "summary": "Revoke refresh token",
        "tags": ["auth"]
      }
    },
    "/api/v1/auth/request-password-reset": {
      "post": {
        "operationId": "AuthController_requestPasswordReset_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequestPasswordResetBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestPasswordResetResponseSchema"
                }
              }
            }
          },
          "400": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsBodySchema"
                }
              }
            }
          },
          "429": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsBodySchema"
                }
              }
            }
          }
        },
        "summary": "Request password reset (neutral response)",
        "tags": ["auth"]
      }
    },
    "/api/v1/auth/resend-verification": {
      "post": {
        "operationId": "AuthController_resendVerification_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateResendVerificationBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResendVerificationResponseSchema"
                }
              }
            }
          },
          "400": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsBodySchema"
                }
              }
            }
          },
          "429": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsBodySchema"
                }
              }
            }
          }
        },
        "summary": "Resend email verification (neutral response)",
        "tags": ["auth"]
      }
    },
    "/api/v1/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateResetPasswordBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResetPasswordResponseSchema"
                }
              }
            }
          },
          "400": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsBodySchema"
                }
              }
            }
          },
          "401": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsBodySchema"
                }
              }
            }
          }
        },
        "summary": "Complete password reset with token",
        "tags": ["auth"]
      }
    },
    "/api/v1/auth/verify-email": {
      "post": {
        "operationId": "AuthController_verifyEmail_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVerifyEmailBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyEmailResponseSchema"
                }
              }
            }
          },
          "400": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsBodySchema"
                }
              }
            }
          },
          "401": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsBodySchema"
                }
              }
            }
          }
        },
        "summary": "Verify email with token",
        "tags": ["auth"]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "AuthController_me_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthMeResponseSchema"
                }
              }
            }
          },
          "401": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsBodySchema"
                }
              }
            }
          },
          "403": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsBodySchema"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Current authenticated user",
        "tags": ["auth"]
      }
    },
    "/api/v1/cms/posts": {
      "get": {
        "operationId": "CmsPostsController_listPosts_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CmsPostsListResponseSchema"
                }
              }
            }
          },
          "401": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsBodySchema"
                }
              }
            }
          },
          "403": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsBodySchema"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "List CMS posts (RBAC probe stub)",
        "tags": ["cms"]
      }
    },
...
  },
  "info": {
    "title": "Blog API",
    "description": "Blog/CMS HTTP API. Versioned routes live under `/api/v1`. Errors use `application/problem+json`.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Access token from login or refresh"
      }
    },
    "schemas": {
      "ProblemDetailsBodySchema": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "https://blog.dev/problems/validation-failed"
          },
          "title": {
            "type": "string",
            "example": "Validation Failed"
          },
          "status": {
            "type": "number",
            "example": 400
          },
          "detail": {
            "type": "string",
            "example": "Request body failed validation."
          },
          "instance": {
            "type": "string",
            "example": "/api/v1/auth/login"
          },
          "code": {
            "type": "string",
            "example": "VALIDATION_FAILED"
          },
          "details": {
            "type": "object",
            "additionalProperties": true,
            "example": {
              "fields": [
                {
                  "path": "email",
                  "message": "email must be an email"
                }
              ]
            }
          }
        },
        "required": ["type", "title", "status", "detail", "code"]
      },
...
      "CreateRegisterBodyDto": {
        "type": "object",
        "properties": {}
      },
      "RegisterUserResponseSchema": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "emailVerificationToken": {
            "type": "string",
            "description": "Omitted when email channel is enabled"
          },
          "emailVerifiedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["id", "email", "emailVerifiedAt", "createdAt", "updatedAt"]
      },
      "CreateLoginBodyDto": {
        "type": "object",
        "properties": {}
      },
      "LoginUserResponseSchema": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "createdAt",
          "updatedAt",
          "accessToken",
          "refreshToken"
        ]
      },
      "CreateRefreshBodyDto": {
        "type": "object",
        "properties": {}
      },
      "RefreshSessionResponseSchema": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          }
        },
        "required": ["accessToken", "refreshToken"]
      },
      "CreateRequestPasswordResetBodyDto": {
        "type": "object",
        "properties": {}
      },
      "RequestPasswordResetResponseSchema": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "If an account exists, further instructions were sent."
          },
          "passwordResetToken": {
            "type": "string",
            "description": "Omitted when email channel is enabled"
          }
        },
        "required": ["message"]
      },
      "CreateResendVerificationBodyDto": {
        "type": "object",
        "properties": {}
      },
      "ResendVerificationResponseSchema": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "If an account exists, further instructions were sent."
          },
          "emailVerificationToken": {
            "type": "string",
            "description": "Omitted when email channel is enabled"
          }
        },
        "required": ["message"]
      },
      "CreateResetPasswordBodyDto": {
        "type": "object",
        "properties": {}
      },
      "ResetPasswordResponseSchema": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Password has been reset."
          }
        },
        "required": ["message"]
      },
      "CreateVerifyEmailBodyDto": {
        "type": "object",
        "properties": {}
      },
      "VerifyEmailResponseSchema": {
        "type": "object",
        "properties": {
          "emailVerifiedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["emailVerifiedAt"]
      },
      "AuthMeResponseSchema": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": ["id"]
      },
      "CmsPostsListResponseSchema": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "example": []
          }
        },
        "required": ["items"]
      },
...
    }
  }
}